requirements.txt
crispy-bootstrap5
Django
django-crispy-forms
keras
Pillow
tensorflow-cpu